requirements.txt
flask
requests
psutil
mss
pillow
pystray
pyperclip
keyboard
pyautogui
python-dotenv
